+++ src/vm/hints/mod.rs
@@ -3,4 +3,4 @@ pub mod dict_manager;
 pub mod execute_hint;
 pub mod hint_utils;
 pub mod pow_utils;
-pub mod squash_dict;
+pub mod squash_dict_utils;
