+++ src/vm/hints/squash_dict_utils.rs
